Book List Addition Functionality and view cleanup

+++ resources/views/books/my_book_lists.blade.php
@@ -1,11 +1,9 @@
-<?php
-$book_lists = App\UserBookList::all();
-?>
 <table class="table table-striped">
     <thead>
     <tr>
         <th>Name</th>
         <th>Description</th>
+        <th>Owner</th>
         <th>Public</th>
         <th>Updated Date</th>
         <th>Created Date</th>
@@ -17,14 +15,19 @@ $book_lists = App\UserBookList::all();
         <tr>
             <td><a href="{{ route('book-list.show', $book_list->id) }}">{{ $book_list->name }}</a></td>
             <td>{{ $book_list->description }}</td>
+            <td>{{ App\User::find($book_list->user_id)->name }}</td>
             <td>{{ ($book_list->public) ? 'Yes' : 'No' }}</td>
             <td>{{ $book_list->updated_at }}</td>
             <td>{{ $book_list->created_at }}</td>
             <td>
-                <a class="btn btn-warning" href="/book-list/{{ $book_list->id }}/edit">Edit</a>
-                {!! Form::open(['method' => 'DELETE','route' => ['book-list.destroy', $book_list->id], 'style'=> 'display: inline']) !!}
-                {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
-                {!! Form::close() !!}
+                @if(Auth::check() == false)
+                    @if($book_list->user_id == Auth::user()->id)
+                    <a class="btn btn-warning" href="/book-list/{{ $book_list->id }}/edit">Edit</a>
+                    {!! Form::open(['method' => 'DELETE','route' => ['book-list.destroy', $book_list->id], 'style'=> 'display: inline']) !!}
+                    {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
+                    {!! Form::close() !!}
+                    @endif
+                @endif
             </td>
         </tr>
     @endforeach
